break up main into separate files
